% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity.R
\name{get_activity}
\alias{get_activity}
\title{Retrieve Mp activity}
\usage{
get_activity(term = "16")
}
\arguments{
\item{term}{Term number}
}
\value{
A tibble
}
\description{
Retrieve Mp activity
}
\examples{
get_activity(term = "16")

}
