% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orga.R
\name{get_orga}
\alias{get_orga}
\title{Retrieve Organisms Data from the French Parliament API}
\usage{
get_orga(term, type = c("groupe"))
}
\arguments{
\item{term}{A character vector specifying the legislature terms. Valid terms are '13', '14', '15', and '16'.}

\item{type}{A character vector specifying the types of the organisms. Default is 'groupe'.}
}
\value{
A tibble containing information about the organisms for the provided terms and types.
}
\description{
This function fetches data about specified organisms (or "orgas") from the French Parliament API
based on given legislature terms and types.
}
\examples{
\dontrun{
terms_list <- c("15", "16")
types_list <- c("groupe", "extra")
all_orga <- get_orga(term = terms_list, type = types_list)
}
}
