% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_vote_mps}
\alias{get_vote_mps}
\title{Get detail of each mp for roll call}
\usage{
get_vote_mps(api_url = NULL, legislature = NULL, vote_number = NULL)
}
\arguments{
\item{api_url}{A character vector indicating the endpoint of the API}

\item{legislature}{The legislature number}

\item{vote_number}{The vote number}
}
\value{
A tibble
}
\description{
Get detail of each mp for roll call
}
\examples{
get_vote_mps(legislature = "16", vote_number = "1")
}
