% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes}
\alias{get_votes}
\title{Get roll call votes}
\usage{
get_votes(term = "16", add_meta = TRUE)
}
\arguments{
\item{term}{A character vector indicating the term}

\item{add_meta}{A boolean indicating if meta data should be added}
}
\value{
A tibble
}
\description{
Get roll call votes
}
\examples{
get_votes("16", add_meta = TRUE)
}
